Support remote files.
When calling the "p4" program, use process functions that can have
handlers for remote file names, and pass those functions the local
version of the file name on the remote host where "p4" will run.

The one difficult function to replace is 'call-process-region', which
has no handled variant 'process-file-region'.  For this we copy a
process-file-region implementation from the 'xref' package.